chore: move assets around
