+++ src/components/banner-stickers/banner-stickers-group.astro
@@ -1,5 +1,5 @@
 ---
-import unicornSticker from "#src/assets/playfulprogramming_sticker.svg?raw";
+import unicornSticker from "#src/assets/logo/playfulprogramming_sticker.svg?raw";
 import { tags } from "#utils/data.ts";
 import styles from "./banner-stickers-group.module.scss";
