+++ src/pages/confirm.astro
@@ -3,7 +3,7 @@ import Document from "#components/layouts/document/document.astro";
 import SEO from "#components/seo/seo.astro";
 import PageCard from "#components/page-card/page-card.astro";
 
-import neutralUnicorn from "../assets/unicorn_neutral.svg";
+import neutralUnicorn from "../assets/emotes/unicorn_neutral.svg";
 import { LargeButton } from "../components/index.ts";
 import { baseLocale } from "#src/paraglide/runtime.js";
