+++ src/pages/thanks.astro
@@ -3,7 +3,7 @@ import Document from "#components/layouts/document/document.astro";
 import SEO from "#components/seo/seo.astro";
 import PageCard from "#components/page-card/page-card.astro";
 
-import happyUnicorn from "../assets/unicorn_happy.svg";
+import happyUnicorn from "../assets/emotes/unicorn_happy.svg";
 import { LargeButton } from "../components/index.ts";
 import { baseLocale } from "#src/paraglide/runtime.js";
