+++ src/utils/markdown/iframes/platform-detectors/post.ts
@@ -3,7 +3,7 @@ import {
 	type ComponentNode,
 	createComponent,
 } from "#utils/markdown/components/index.ts";
-import unicornHappy from "#src/assets/unicorn_happy.svg?url";
+import unicornHappy from "#src/assets/emotes/unicorn_happy.svg?url";
 
 export function rehypeTransformPost({
 	src,
