+++ src/views/about/banner-svg.astro
@@ -3,7 +3,7 @@ import fs from "fs/promises";
 import { getPeopleByLang } from "#utils/api.ts";
 import style from "./banner-svg.module.scss";
 import { Picture } from "#components/image/picture.tsx";
-import logo from "#src/assets/playfulprogramming_circle.svg?svg?raw";
+import logo from "#src/assets/logo/playfulprogramming_circle.svg?svg?raw";
 import { m } from "#src/paraglide/messages.js";
 import { baseLocale } from "#src/paraglide/runtime.js";
