+++ src/views/base/navigation/header.astro
@@ -6,7 +6,7 @@ import BackButton from "./back-button/back-button.astro";
 import ThemeSidebar from "./theme-sidebar/theme-sidebar.astro";
 import ThemeSidebarTrigger from "./theme-sidebar/theme-sidebar-trigger.astro";
 import style from "./header.module.scss";
-import logo from "#src/assets/playfulprogramming_sticker.svg?raw";
+import logo from "#src/assets/logo/playfulprogramming_sticker.svg?raw";
 import { m } from "#src/paraglide/messages.js";
 import { SearchInput } from "#components/input/input.tsx";
 import { SEARCH_QUERY_KEY } from "#src/views/search/search.ts";
