+++ src/views/not-found/not-found-page.astro
@@ -2,7 +2,7 @@
 import Document from "#components/layouts/document/document.astro";
 import SEO from "#components/seo/seo.astro";
 import PageCard from "#components/page-card/page-card.astro";
-import sadUnicorn from "#src/assets/unicorn_sad.svg";
+import sadUnicorn from "#src/assets/emotes/unicorn_sad.svg";
 import { LargeButton } from "#components/index.ts";
 import { getLocale, localizeHref } from "#src/paraglide/runtime.js";
 import { m } from "#src/paraglide/messages.js";
