+++ src/views/search/search-page.tsx
@@ -24,9 +24,9 @@ import { SearchTopbar } from "./components/search-topbar.tsx";
 import { SearchHero } from "./components/search-hero.tsx";
 import { LargeButton } from "#components/button/button.tsx";
 import retry from "#src/assets/icons/refresh.svg?raw";
-import sadUnicorn from "../../assets/unicorn_sad.svg";
-import happyUnicorn from "../../assets/unicorn_happy.svg";
-import scaredUnicorn from "../../assets/unicorn_scared.svg";
+import sadUnicorn from "../../assets/emotes/unicorn_sad.svg";
+import happyUnicorn from "../../assets/emotes/unicorn_happy.svg";
+import scaredUnicorn from "../../assets/emotes/unicorn_scared.svg";
 import {
 	type SearchQuery,
 	type DisplayContentType,
